fix Nose being imported outside tests
