+++ hdbscan/__init__.py
@@ -2,6 +2,5 @@ from .hdbscan_ import HDBSCAN, hdbscan
 from .robust_single_linkage_ import RobustSingleLinkage, robust_single_linkage
 from .validity import validity_index
 from .prediction import approximate_predict, membership_vector, all_points_membership_vectors
-from .utils import if_matplotlib
